requirement.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib
bcrypt==3.2.2
pytest
python-jose[cryptography]
email-validator
python-multipart
httpx
